requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
google-generativeai>=0.7.0
pinecone>=5.0.0
numpy>=1.26.0
pydantic>=2.9.0
python-multipart>=0.0.9
